Modularize legacy import pipeline
Separate legacy row contracts, run-ledger persistence, conflict policy, row application, and behavior tests behind the existing db-app import API.
